GH-61 product 테스트 코드 개선
