+++ src/main/java/com/example/study/product/command/domain/product/ProductRepository.java
@@ -1,38 +1,27 @@
 package com.example.study.product.command.domain.product;
 
 import jakarta.persistence.LockModeType;
-import jakarta.persistence.QueryHint;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Lock;
-import org.springframework.data.jpa.repository.QueryHints;
 
 import java.util.List;
 import java.util.Optional;
 
+/**
+ * 비관락이므로 서비스 단계에서 @DBLockTimeout 이 같이 선언되어야 한다.
+ */
 public interface ProductRepository extends JpaRepository<Product, Long> {
 
     @Lock(LockModeType.PESSIMISTIC_WRITE)
-    @QueryHints({
-            @QueryHint(name = "javax.persistence.lock.timeout", value = "3000") // 3초 대기
-    })
     Optional<DeliveryProduct> findDeliveryProductById(Long id);
 
     @Lock(LockModeType.PESSIMISTIC_WRITE)
-    @QueryHints({
-            @QueryHint(name = "javax.persistence.lock.timeout", value = "3000") // 3초 대기
-    })
     Optional<CouponProduct> findCouponProductById(Long id);
 
     @Lock(LockModeType.PESSIMISTIC_WRITE)
-    @QueryHints({
-            @QueryHint(name = "javax.persistence.lock.timeout", value = "3000") // 3초 대기
-    })
     Optional<Product> findProductById(Long id);
 
     @Lock(LockModeType.PESSIMISTIC_WRITE)
-    @QueryHints({
-            @QueryHint(name = "javax.persistence.lock.timeout", value = "3000") // 3초 대기
-    })
     List<Product> findAllByIdIn(List<Long> ids);
 
 }
